requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
boto3==1.28.3
botocore==1.31.3
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
Flask==2.3.2
Flask-Cors==4.0.0
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
load-dotenv==0.1.0
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.8
packaging==23.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
tqdm==4.65.0
urllib3==1.26.16
Werkzeug==2.3.6
yarl==1.9.2
